% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/can_IEEE_connect.R
\name{can_IEEE_connect}
\alias{can_IEEE_connect}
\title{Check for connection to IEEE API}
\usage{
can_IEEE_connect(max_time = 5)
}
\arguments{
\item{max_time}{Maximum wait time in seconds}
}
\value{
Returns TRUE if connection is established and FALSE
otherwise.
}
\description{
Check for connection to IEEE API
}
\examples{
\donttest{
can_IEEE_connect(2)
}
}

