% Generated by roxygen2 (4.0.2): do not edit by hand
\name{IEEE_count}
\alias{IEEE_count}
\title{Count number of results for a given search}
\usage{
IEEE_count(query = NULL)
}
\arguments{
\item{query}{Character string (for a simple search), or a named
list of search strings, with the names corresponding to IEEE search
parameters (see \code{\link{query_param}} or
\url{http://ieeexplore.ieee.org/gateway})}
}
\value{
Number of results (integer). An attribute
\code{"search_info"} contains information about the search
parameters and the time at which it was performed.
}
\description{
Count the number of results for a given search. Useful to check
before attempting to pull down a very large number of records.
}
\examples{
\dontshow{old_delay <- getOption("IEEER_delay")
          options(IEEER_delay=1)}
# count papers in category stat.AP (applied statistics)
IEEE_count(query = "cat:stat.AP")

# count papers by Peter Hall in any stat category
\donttest{IEEE_count(query = 'au:"Peter Hall" AND cat:stat*')}

# count papers for a range of dates
#    here, everything in 2013
\donttest{IEEE_count("submittedDate:[2013 TO 2014]")}
\dontshow{options(IEEER_delay=old_delay)}
}

