% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{query_param}
\alias{query_param}
\title{IEEE query field parameters}
\format{A data frame with four columns: the \code{term}, \code{description}, \code{boolean_query_field},
and \code{type}.}
\source{
\url{http://ieeexplore.ieee.org/gateway}
}
\usage{
data(query_param)
}
\description{
Possible parameters for IEEE searches.
}
\examples{
query_param[1:10,1:2]
}
\keyword{datasets}

