% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IEEE_count.R
\name{IEEE_count}
\alias{IEEE_count}
\title{Count number of results for a given search}
\usage{
IEEE_count(query = NULL)
}
\arguments{
\item{query}{Character string (for a simple search), or a named
list of search strings, with the names corresponding to IEEE search
parameters (see \code{\link{query_param}} or
\url{http://ieeexplore.ieee.org/gateway})}
}
\value{
Number of results (integer). An attribute
\code{"search_info"} contains information about the search
parameters and the time at which it was performed.
}
\description{
Count the number of results for a given search. Useful to check
before attempting to pull down a very large number of records.
}
\examples{
\donttest{
# count papers in year 1970
IEEE_count(list(pys=1970, pye=1970))

# count papers by author Rabiner
IEEE_count(list(au="Rabiner"))

# count papers by author Rabiner, L
IEEE_count(list(au="Rabiner, L"))
}
}
