% Generated by roxygen2 (4.0.2): do not edit by hand
\name{IEEE_search}
\alias{IEEE_search}
\title{The main search function for IEEE}
\usage{
IEEE_search(query = NULL, start = 0, limit = 10, sort_by = c("year",
  "author", "title", "affilitation", "journal", "articlenumber"),
  ascending = TRUE, batchsize = 100, output_format = c("data.frame",
  "list"), sep = "|")
}
\arguments{
\item{query}{Character string (for a simple search), or a named
list of search strings, with the names corresponding to IEEE search
parameters (see \code{\link{query_param}} or
\url{http://ieeexplore.ieee.org/gateway})}

\item{start}{An offset for the start of search}

\item{limit}{Maximum number of records to return}

\item{sort_by}{How to sort the results}

\item{ascending}{If TRUE, sort in ascending order; else descending}

\item{batchsize}{Maximum number of records to request at one time}

\item{output_format}{Indicates whether output should be a data frame or a list}

\item{sep}{String to use to separate multiple entries in
multiple-entry fields (e.g., \code{controlledterms}), in the case
that \code{output_format="data.frame"}.}
}
\value{
If \code{output_format="data.frame"}, the result is a data
frame with each row being a manuscript and columns being the
various fields.

If \code{output_format="list"}, the result is a list parsed from
the XML output of the search, closer to the raw output from IEEE.

The data frame format has the following columns.
\tabular{rll}{
 [,1] \tab rank             \tab numeric rank in search \cr
 [,2] \tab title            \tab document title \cr
 [,3] \tab authors          \tab authors \cr
 [,4] \tab thesaurusterms   \tab thesaurus terms \cr
 [,5] \tab pubtitle         \tab journal title \cr
 [,6] \tab punumber         \tab publication number \cr
 [,7] \tab pubtype          \tab publication type \cr
 [,8] \tab publisher        \tab publisher \cr
 [,9] \tab py               \tab publication year \cr
[,10] \tab spage            \tab start page \cr
[,11] \tab epage            \tab end page \cr
[,12] \tab abstract         \tab abstract \cr
[,13] \tab isbn             \tab ISBN number \cr
[,14] \tab htmlFlag         \tab html flag (\code{"1"} or \code{""}) \cr
[,15] \tab arnumber         \tab article number \cr
[,16] \tab doi              \tab DOI \cr
[,17] \tab publicationId    \tab publication ID \cr
[,18] \tab mdurl            \tab document URL \cr
[,19] \tab pdf              \tab URL for PDF \cr
[,20] \tab affiliations     \tab authors' affiliations \cr
[,21] \tab controlledterms  \tab controlled terms (keywords) \cr
[,22] \tab volume           \tab volumn number \cr
[,23] \tab issn             \tab ISSN number \cr
[,24] \tab issue            \tab issue number \cr
}

The contents are all strings; missing values are empty strings (\code{""}).

Some columns (e.g., \code{thesaurusterms} and
\code{controlledterms}) may have multiple entries separated by
\code{sep} (by default, \code{"|"}).

The result includes an attribute \code{"search_info"} that includes
information about the details of the search parameters, including
the time at which it was completed. Additional attributes include
\code{"totalfound"}, the total number of records that match the
query, and \code{totalsearched}, the total number of records
searched.
}
\description{
Allows for progammatic searching of the IEEE pre-print repository.
}
\examples{
\dontshow{old_delay <- getOption("IEEER_delay")
          options(IEEER_delay=1)}
# search for author Peter Hall with deconvolution in title
z <- IEEE_search(query = 'au:"Peter Hall" AND ti:deconvolution', limit=2)
attr(z, "total_results") # total no. records matching query
z$title

# search for a set of documents by IEEE identifiers
\donttest{z <- IEEE_search(id_list = c("0710.3491v1", "0804.0713v1", "1003.0315v1"))}
# can also use a comma-separated string
\donttest{z <- IEEE_search(id_list = "0710.3491v1,0804.0713v1,1003.0315v1")}
# Journal references, if available
z$journal_ref

# search for a range of dates (in this case, one day)
\donttest{z <- IEEE_search("submittedDate:[199701010000 TO 199701012400]", limit=2)}
\dontshow{options(IEEER_delay=old_delay)}
}

