% Generated by roxygen2 (4.0.2): do not edit by hand
\name{IEEE_search}
\alias{IEEE_search}
\title{The main search function for IEEE}
\usage{
IEEE_search(query = NULL, start = 1, limit = 10, sort_by = c("year",
  "author", "title", "affilitation", "journal", "articlenumber"),
  ascending = TRUE, batchsize = 100, output_format = c("data.frame",
  "list"), sep = "|")
}
\arguments{
\item{query}{Character string (for a simple search), or a named
list of search strings, with the names corresponding to IEEE search
parameters (see \code{\link{query_param}} or
\url{http://ieeexplore.ieee.org/gateway})}

\item{start}{An offset for the start of search (\code{>= 1})}

\item{limit}{Maximum number of records to return}

\item{sort_by}{How to sort the results}

\item{ascending}{If TRUE, sort in ascending order; else descending}

\item{batchsize}{Maximum number of records to request at one time}

\item{output_format}{Indicates whether output should be a data frame or a list}

\item{sep}{String to use to separate multiple entries in
multiple-entry fields (e.g., \code{controlledterms}), in the case
that \code{output_format="data.frame"}.}
}
\value{
If \code{output_format="data.frame"}, the result is a data
frame with each row being a manuscript and columns being the
various fields.

If \code{output_format="list"}, the result is a list parsed from
the XML output of the search, closer to the raw output from IEEE.

The data frame format has the following columns.
\tabular{rll}{
 [,1] \tab rank              \tab numeric rank in search \cr
 [,2] \tab title             \tab article title \cr
 [,3] \tab authors           \tab semicolon delimited list of author names \cr
 [,4] \tab affiliations      \tab authors' affiliations \cr
 [,5] \tab pubtitle          \tab publication in which article appears \cr
 [,6] \tab punumber          \tab IEEE identifier for publication \cr
 [,7] \tab py                \tab publication year \cr
 [,8] \tab volume            \tab volume number \cr
 [,9] \tab issue             \tab issue number \cr
[,10] \tab part              \tab part \cr
[,11] \tab spage             \tab start page \cr
[,12] \tab epage             \tab end page \cr
[,13] \tab arnumber          \tab unique article number \cr
[,14] \tab abstract          \tab first 250 words of the abstract \cr
[,15] \tab doi               \tab digital object identifier \cr
[,16] \tab mdurl             \tab document URL \cr
[,17] \tab pdf               \tab URL for PDF \cr
[,18] \tab pubtype           \tab publication type (Journal, Conference, or Standard)\cr
[,19] \tab publisher         \tab publisher \cr
[,20] \tab isbn              \tab ISBN number \cr
[,21] \tab issn              \tab ISSN number \cr
[,22] \tab publicationId     \tab publication ID \cr
[,23] \tab thesaurusterms    \tab terms from IEEE thesaurus \cr
[,24] \tab controlledterms   \tab terms from INSPEC controlled thesaurus \cr
[,25] \tab uncontrolledterms \tab terms not from INSPEC thesaurus \cr
[,26] \tab htmlFlag          \tab html flag (\code{"1"} or \code{""}) \cr
}

The contents are all strings; missing values are empty strings (\code{""}).

Some columns (e.g., \code{thesaurusterms} and
\code{controlledterms}) may have multiple entries separated by
\code{sep} (by default, \code{"|"}).

The result includes an attribute \code{"search_info"} that includes
information about the details of the search parameters, including
the time at which it was completed. Additional attributes include
\code{"totalfound"}, the total number of records that match the
query, and \code{totalsearched}, the total number of records
searched.
}
\description{
Allows for progammatic searching of the IEEE pre-print repository.
}
\examples{
# search for author Rabiner with Markov in title
z <- IEEE_search(list(au="Rabiner", ti="Markov"), limit=2)
attr(z, "totalfound") # total no. records matching query
z$title

# search for author Rabiner in years 1960-1970
\donttest{z <- IEEE_search(list(au="Rabiner", pys=1960, pye=1970))}
}

